apply titlecasing on NameNetwork when case change count is 0
