+++ lib/cluster.js
@@ -9,13 +9,19 @@ const simple = require('simple-statistics');
 const Misc = require('./misc');
 const misc = new Misc();
 const linker = require('./linker');
+const numCaseChanges = require('./label/util').numCaseChanges;
+
 
 /**
  * @class Cluster
  */
 class Cluster {
-    constructor(pool) {
-        this.pool = pool;
+    constructor(opts) {
+        this.pool = opts.pool;
+        if (opts.label)
+            this.label = require('./label/' + opts.label)();
+        else
+            this.label = require('./label/titlecase')({ language: 'en' });
     }
 
     break(segs) {
@@ -842,6 +848,9 @@ class Cluster {
                         } else {
                             // any string needs to be made sql safe: replace single quote with two single quotes
                             new_text_values[k] = new_text_values[k].replace(/'/g, "''");
+                            // titlecase / run label func if zero case changes
+                            if (numCaseChanges(new_text_values[k]) === 0)
+                                new_text_values[k] = this.label(new_text_values[k]);
                         }
                     }
