+++ lib/label/titlecase.js
@@ -4,33 +4,8 @@ const wordBoundary = new RegExp("[\\s\\u2000-\\u206F\\u2E00-\\u2E7F\\\\!\"#$%&()
 const allowedAfterSpace = new RegExp("[()#<>\\[\\]{}\"]+");
 const diacritics = require('diacritics').remove;
 const dist = require('fast-levenshtein').get
-
-/**
- * Test if a string if fully uppercase
- * @param {string} x test input
- * @return {boolean}
- */
-function isUpperCase(x) {
-    return x.toUpperCase() === x;
-}
-
-/**
- * Number of times a string switches from lower to upper or upper to lower
- * @param {string} x String to test
- * @return {numeric}
- */
-function numCaseChanges(x) {
-    return x
-        .split('')
-        .filter((y) => { return /\w/.test(y); })
-        .reduce((prev, cur, i, arr) => {
-            if (i === 0) return 0;
-            if (isUpperCase(cur) !== isUpperCase(arr[i-1]))
-                return prev + 1;
-            else
-                return prev;
-        }, 0);
-}
+const numCaseChanges = require('./util').numCaseChanges;
+const isUpperCase = require('./util').isUpperCase;
 
 /**
  * Title case a given string except for minors
