+++ lib/map.js
@@ -104,7 +104,9 @@ function main(argv, cb) {
 
     const pool = new pg.Pool(poolConf);
 
-    const cluster = new Cluster(pool);
+    let opts = { pool: pool };
+    if (argv.label) opts.label = argv.label;
+    const cluster = new Cluster(opts);
     const index = new Index(pool);
     const orphan = new Orphan(pool, argv, output, post);
