+++ lib/orphan.js
@@ -33,7 +33,7 @@ class Orphan {
     address(cb) {
         const self = this;
 
-        const cluster = new Cluster(this.pool);
+        const cluster = new Cluster({ pool: this.pool });
 
         cluster.orphan((err) => {
