+++ lib/split.js
@@ -77,7 +77,7 @@ function init(o) {
 
     pool = new pg.Pool(opts.pool);
 
-    cluster = new Cluster(pool);
+    cluster = new Cluster({ pool: pool });
 
     if (!opts.label) {
         label = require('./label/titlecase')({ language: 'en', synonym: true });
